# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/product/create": {
      "post": {
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建商品",
        "tags": [
          "product"
        ]
      }
    },
    "/api/product/list": {
      "get": {
        "operationId": "ProductController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "商品名称模糊搜索",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "商品状态筛选",
            "schema": {
              "type": "number",
              "enum": [
                0,
                1,
                2
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "商品列表（分页 + 名称模糊 + 状态筛选）",
        "tags": [
          "product"
        ]
      }
    },
    "/api/product/hot-list": {
      "get": {
        "operationId": "ProductController_hotList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "热销商品",
        "tags": [
          "product"
        ]
      }
    },
    "/api/product/edit": {
      "patch": {
        "operationId": "ProductController_edit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "编辑商品（上下架走 updateStatus）",
        "tags": [
          "product"
        ]
      }
    },
    "/api/product/updateStatus": {
      "patch": {
        "operationId": "ProductController_updateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "商品上下架",
        "tags": [
          "product"
        ]
      }
    },
    "/api/product/import": {
      "post": {
        "operationId": "ProductController_importProducts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultVo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "excel 批量导入商品（表头：商品名称 | 商品描述 | 价格）",
        "tags": [
          "product"
        ]
      }
    },
    "/api/product/{id}": {
      "delete": {
        "operationId": "ProductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除商品（被订单/活动引用时拒删）",
        "tags": [
          "product"
        ]
      }
    },
...
  },
  "info": {
    "title": "store-web API",
    "description": "数字门店系统后端接口",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "商品名称",
            "example": "烧鹅"
          },
          "desc": {
            "type": "string",
            "description": "商品描述",
            "example": "今日现烧"
          },
          "price": {
            "type": "number",
            "description": "商品价格",
            "example": 38.5
          },
          "images": {
            "type": "string",
            "description": "商品图片"
          },
          "status": {
            "type": "number",
            "description": "商品状态 0 未上架 1 已上架 2 已下架",
            "enum": [
              0,
              1,
              2
            ]
          }
        },
        "required": [
          "name",
          "desc",
          "price"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "商品 id"
          },
          "name": {
            "type": "string",
            "description": "商品名称"
          },
          "images": {
            "type": "string",
            "description": "商品图片",
            "nullable": true
          },
          "desc": {
            "type": "string",
            "description": "商品描述"
          },
          "status": {
            "type": "number",
            "description": "商品状态 0 未上架 1 已上架 2 已下架"
          },
          "price": {
            "type": "number",
            "description": "商品价格"
          },
          "createTime": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updateTime": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "name",
          "images",
          "desc",
          "status",
          "price",
          "createTime",
          "updateTime"
        ]
      },
      "ProductListVo": {
        "type": "object",
        "properties": {
          "list": {
            "description": "当前页数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "total": {
            "type": "number",
            "description": "总条数"
          }
        },
        "required": [
          "list",
          "total"
        ]
      },
      "EditProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "商品 id"
          },
          "name": {
            "type": "string",
            "description": "商品名称"
          },
          "desc": {
            "type": "string",
            "description": "商品描述"
          },
          "price": {
            "type": "number",
            "description": "商品价格"
          },
          "images": {
            "type": "string",
            "description": "商品图片"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateProductStatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "商品 id"
          },
          "status": {
            "type": "number",
            "description": "商品状态 0 未上架 1 已上架 2 已下架",
            "enum": [
              0,
              1,
              2
            ]
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
      "ImportResultVo": {
        "type": "object",
        "properties": {
          "imported": {
            "type": "number",
            "description": "成功导入条数"
          }
        },
        "required": [
          "imported"
        ]
      },
...
    }
  }
}